Add base configuration
This commit includes a vagrant file which:

1. Creates a Debian virtual machine
2. Mounts torque-sites (your working copy)
3. Runs the three "base" ansible scripts (mwlib, torquedata, simplesaml)

It is possible that in future we will want the torquedata code to be
a working copy as well, but for now this should do the trick.

This script will NOT set up any competitions / instances of
torque-sites, but it does create a baseline machine that we can modify
to accomplish that later.

All of this is built from the AMAZING work of Cristina Muñoz at https://github.com/PermanentOrg/devenv

Issue #1