Move to dhcp
We were originally using static IP configuration, but decided to move to
dhcp on the off chance that the static IP we picked already existed in
a given network.  This becomes less predictable, and means you can't
have a local hostfile, but it seems reasonable for now.

If hostfiles become important there are vagrant plugins to help manage
that.