Enable ETL
This commit enables support for automated setup and execution of ETL
processes for competitions. This means it allows for automatic
downloading of data and utilities from an OTS (or OTS compatable) SVN
repository as well as the necessary tools to eventually run the `deploy`
scripts hosted in torque-sites.

The new environment variables that are documented in the installation
steps / template files are optional, and if they are ignored then setup
will still work but will not support ETL.

The SSL cert of the OTS subversion repository is included here, which
allows us to securely (and automatically) interact with that server
without being forced to ignore certificates.

Issue #3