Override default Apache configuration
The default Apache configuration does not allow for htaccess overrides.
Some competitions need this to function (particularly when it comes to
the default upload size).

The 100Change competition is one such competition.

Without this change, it's not possible to run the ETL pipeline because
apache rejects the csv upload for being too large