Add SimpleBook support
SimpleBook is a wikimedia extension being developed as part of this
project.  It is part of the torque ecosystem, and so the torque devenv
should support development of it.

SimpleBook relies on both nodejs and Python.  It also uses a headless
chromium package called pippeteer with its own set of system
requirements.

Issue #8