Install SimpleBook via Ansible
We no longer want to install SimpleBook "manually" but instead want to
use the Ansible scripts that are part of torque-sites.

This temporarily removes SimpleBook from devenv (as in, the mounted copy
is not actually used in any local competitions), but it does make sure
our local setup is performed in the same way it will be performed on
production.

The next step will be to link in the mounted code in the various
directories that we want to use it.

Since we're relying on Ansible to install SimpleBook we don't (yet) need
to install pipenv or node, so those dependencies are removed for now.

Issue #20