Use mounted SimpleBook
We want to install SimpleBook using ansible, but we also want to use our
local copy since this is, at the end of the day, the development
environment.

To do this we delete the downloaded copies of SimpleBook and instead
symlink our mounted copy.  We also need to install the SimpleBook
dependencies (e.g. pipenv and yarn install) so that we can set up our
mounted copy of SimpleBook properly.

Technically those tools are installed via Ansible but in the spirit of
defensiveness we'll explicitly install them in our provisioner as well
in case that ever changes / to ensure that the development requirements
of SimpleBook are reflected in the development environment.

Issue #20