Add SimpleMaps
This adds SimpleMaps to the development environment, though it does not
install the local copy to any competitions (this will still need to be
done manually for the competition in question).

You can "add" it to a competition by creating a link between the
competition's directory and the ~/SimpleMaps directory.

Issue #30