Save environment variables
Some of the environment variables used in the provisioning script are
used by things like ETL and ansible, which are manually run within the
guest machine (outside of provisioning).

By adding these values to the .profile of the vagrant user we prevent
the dev from having to manually export themselves before running those
scripts.

Issue #38